requirements.txt
asgiref==3.4.1
Django==3.2.9
djangorestframework==3.12.4
olefile==0.46
Pillow==8.4.0
pip==21.3.1
pyodbc==4.0.32
pytz==2021.3
setuptools==59.2.0
sqlparse==0.4.2
typing_extensions==4.0.0
wheel==0.37.0
